+++ jupyter_client/tests/test_kernelmanager.py
@@ -21,6 +21,7 @@ from jupyter_client import KernelManager, AsyncKernelManager
 from subprocess import PIPE
 
 from ..manager import start_new_kernel, start_new_async_kernel
+from ..manager import _ShutdownStatus
 from .utils import test_env, skip_win32, AsyncKernelManagerSubclass, AsyncKernelManagerWithCleanup
 
 pjoin = os.path.join
@@ -52,18 +53,47 @@ def config(transport):
     return c
 
 
+def _install_kernel(name, extra_env=None):
+    if extra_env is None:
+        extra_env == {}
+    kernel_dir = pjoin(paths.jupyter_data_dir(), "kernels", name)
+    os.makedirs(kernel_dir)
+    with open(pjoin(kernel_dir, "kernel.json"), "w") as f:
+        f.write(
+            json.dumps(
+                {
+                    "argv": [
+                        sys.executable,
+                        "-m",
+                        "jupyter_client.tests.signalkernel",
+                        "-f",
+                        "{connection_file}",
+                    ],
+                    "display_name": "Signal Test Kernel",
+                    "env": {
+                        "TEST_VARS": "${TEST_VARS}:test_var_2",
+                        "NO_SHUTDOWN_REPLY": "1",
+                    },
+                }
+            )
+        )
+
+
 @pytest.fixture
 def install_kernel():
-    kernel_dir = pjoin(paths.jupyter_data_dir(), 'kernels', 'signaltest')
-    os.makedirs(kernel_dir)
-    with open(pjoin(kernel_dir, 'kernel.json'), 'w') as f:
-        f.write(json.dumps({
-            'argv': [sys.executable,
-                     '-m', 'jupyter_client.tests.signalkernel',
-                     '-f', '{connection_file}'],
-            'display_name': "Signal Test Kernel",
-            'env': {'TEST_VARS': '${TEST_VARS}:test_var_2'},
-        }))
+    return _install_kernel("signaltest")
+
+
+@pytest.fixture
+def install_kernel_dont_shutdown():
+    _install_kernel("signaltest-no-shutdown", {"NO_SHUTDOWN_REPLY": "1"})
+
+
+@pytest.fixture
+def install_kernel_dont_terminate():
+    return _install_kernel(
+        "signaltest-no-terminate", {"NO_SHUTDOWN_REPLY": "1", "NO_SIGTERM_REPLY": "1"}
+    )
 
 
 @pytest.fixture
@@ -103,6 +133,36 @@ async def start_async_kernel():
     kc.stop_channels()
     await km.shutdown_kernel()
     assert km.context.closed
+
+
+class TestKernelManagerShutDownGracefully:
+    @pytest.mark.skipif(
+        sys.platform == "win32", reason="Windows doesn't support signals"
+    )
+    @pytest.mark.parametrize(
+        "name, install, expected",
+        [
+            ("signaltest", install_kernel, _ShutdownStatus.ShutdownRequest),
+            # (
+            #    "signaltest-no-shutdown",
+            #    install_kernel_dont_shutdown,
+            #    _ShutdownStatus.SigtermRequest,
+            # ),
+            # (
+            #    "signaltest-no-terminate",
+            #    install_kernel_dont_terminate,
+            #    _ShutdownStatus.SigKillRequest,
+            # ),
+        ],
+    )
+    def test_signal_kernel_subprocesses(self, name, install, expected):
+        km, kc = start_new_kernel(kernel_name=name)
+        assert km.is_alive()
+        # kc.execute("1")
+        kc.stop_channels()
+        km.shutdown_kernel()
+
+        assert km._shutdown_status == expected
 
 
 class TestKernelManager: